tweaks

+++ web/concrete/tools/conversations/delete_file.php
@@ -1,5 +1,5 @@
 <?php defined('C5_EXECUTE') or die("Access Denied.");
-use \Concrete\Core\Conversation\Message as ConversationMessage;
+use \Concrete\Core\Conversation\Message\Message as ConversationMessage;
 $ax = Loader::helper('ajax');
 $vs = Loader::helper('validation/strings');
 $ve = Loader::helper('validation/error');
@@ -8,10 +8,13 @@ if (Loader::helper('validation/numbers')->integer($_POST['cnvMessageAttachmentID
 	$attachment = ConversationMessage::getAttachmentByID($_POST['cnvMessageAttachmentID']);
 	
 	$message = ConversationMessage::getByID($attachment->cnvMessageID);
-	if (is_object($attachment)) {
-		$message->removeFile($_POST['cnvMessageAttachmentID']);
-	}
-	$attachmentDeleted = new stdClass();
-	$attachmentDeleted->attachmentID = $_POST['cnvMessageAttachmentID'];
-	echo Loader::helper('json')->encode($attachmentDeleted);
+    $mp = new Permissions($message);
+    if ($mp->canEditConversationMessage()) {
+        if (is_object($attachment)) {
+            $message->removeFile($_POST['cnvMessageAttachmentID']);
+        }
+        $attachmentDeleted = new stdClass();
+        $attachmentDeleted->attachmentID = $_POST['cnvMessageAttachmentID'];
+        echo Loader::helper('json')->encode($attachmentDeleted);
+    }
 }
